Create Optimal_partition_of_string.cpp
In the given code, c is a variable that is used in a for loop to iterate over the characters of the input string s.

For each character c in the input string s, the code checks if c is already present in a set ss. If c is present in the set, it means that the current substring has already been seen before and a new substring needs to be started. Therefore, the variable ans is incremented by 1 and the set ss is reset to contain only the current character c.

If c is not present in the set ss, it is added to the set and the loop continues to the next character in the input string s.

The overall goal of the function partitionString is to divide the input string s into the minimum number of substrings such that each character appears in only one substring. The variable c is used in the for loop to achieve this goal.